made edit species and spawnrate, made delete functionality

+++ backend-app3/src/main/java/com/cs440/backend/backendapp3/objects/SpeciesAndSpawnRate.java
@@ -3,14 +3,14 @@ package com.cs440.backend.backendapp3.objects;
 public class SpeciesAndSpawnRate {
 
 	private Species species;
-	private int spawnRate;
+	private float spawnRate;
 	
 	public SpeciesAndSpawnRate() {
 		species = new Species();
 		spawnRate = 0;
 	}
 	
-	public SpeciesAndSpawnRate(int pokedexNum, String name, String type1, String type2, int evolutionNum, int spawnRate) {
+	public SpeciesAndSpawnRate(int pokedexNum, String name, String type1, String type2, int evolutionNum, float spawnRate) {
 		this.species = new Species(pokedexNum, name, type1, type2, evolutionNum);
 		this.spawnRate = spawnRate;
 	}
@@ -21,10 +21,10 @@ public class SpeciesAndSpawnRate {
 	public void setSpecies(Species species) {
 		this.species = species;
 	}
-	public int getSpawnRate() {
+	public float getSpawnRate() {
 		return spawnRate;
 	}
-	public void setSpawnRate(int spawnRate) {
+	public void setSpawnRate(float spawnRate) {
 		this.spawnRate = spawnRate;
 	}
